temporary fix for segv in ZLGtkSignalUtil::removeAllSignals... It seems correct to increment the reference count when adding a GObject to a plain Cpp list

+++ zlibrary/ui/src/gtk/util/ZLGtkSignalUtil.h
@@ -33,7 +33,7 @@ public:
 	static void removeAllSignals();
 
 private:
-	static std::vector<std::pair<GObject*,int> > ourConnectedSignals;
+	static std::vector<std::pair<GObject*, gulong> > ourConnectedSignals;
 
 private:
 	ZLGtkSignalUtil();
